enhance(deploy) made script generic

+++ deploy.sh
@@ -1,6 +1,19 @@
-# The location of your server access key
-key=~/keys/leesjensen/production.pem
-hostname=demo.cs260.click
+while getopts k:h:s: flag
+do
+    case "${flag}" in
+        k) key=${OPTARG};;
+        h) hostname=${OPTARG};;
+        s) service=${OPTARG};;
+    esac
+done
+
+if [[ -z "$key" || -z "$hostname" || -z "$service" ]]; then
+    printf "\nMissing required parameter.\n"
+    printf "  syntax: deploy -k <pem key> -h <hostname> -s <service>\n\n"
+    exit 1
+fi
+
+printf "\n//------------\n//Deploying $service to $hostname with $key\n//------------\n\n"
 
 # Build the distribution package
 rm -rf dist
@@ -9,21 +22,22 @@ cp -r application dist
 cp index.js dist
 cp package* dist
 
-# Clear out the previous distribution
-ssh -i $key ubuntu@$hostname <<'ENDSSH'
-rm -rf services/simon-server
-mkdir -p services/simon-server
+# Clear out the previous distribution on the target.
+ssh -i $key ubuntu@$hostname << ENDSSH
+rm -rf services/${service}
+mkdir -p services/${service}
 ENDSSH
 
-# Copy the distribution package to the server
-scp -r -i $key dist/* ubuntu@$hostname:services/simon-server
+# Copy the distribution package to the target.
+scp -r -i $key dist/* ubuntu@$hostname:services/$service
 
-# Install the package. Install node modules here so they are correct for the server.
-ssh -i $key ubuntu@$hostname <<'ENDSSH'
-cd services/simon-server
+# Install the package on the target.
+# Install node modules here so they are correct for the target.
+ssh -i $key ubuntu@$hostname << ENDSSH
+cd services/${service}
 npm install
-pm2 restart simon-server
+pm2 restart ${service}
 ENDSSH
 
-# Delete the local copy of the package
+# Delete the local copy of the package.
 rm -rf dist
